requirements.txt
numpy
requests
pandas
python-dotenv
sentence-transformers


